+++ src/utils.rs
@@ -1,9 +1,2 @@
-#[cfg(feature = "asyncify")]
-pub mod asyncify;
 pub mod http;
 pub mod io;
-pub mod mutex;
-#[cfg(feature = "atomic-waker")]
-pub mod notification;
-#[cfg(all(feature = "alloc", feature = "atomic-waker"))]
-pub mod zerocopy;